Added location numbers, modified path text.
Added location numbers to the drop down list to make it easier to navigate. Made the output a label to conserve space by making use of word wrapping. Found a bug where choosing a location for the destination would delete it from the source permanently. Disabled the deleting for now.